+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Unreleased
 
+- `CNG` Rejected messages are removed from the approval section
 - `CNG` Improve message listings
 - `ADD` Messages can be deleted